# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "taskcube",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "v1.0"
    },
    "host": "localhost:8080",
    "basePath": "/taskcube/api/v1",
    "paths": {
...
        "/script": {
            "get": {
                "description": "查询执行脚本",
                "tags": [
                    "执行脚本"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "pageSize",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "version",
                        "name": "version",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "shell",
                            "python"
                        ],
                        "type": "string",
                        "description": "Script type",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/services.ScriptService"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "更新执行脚本",
                "tags": [
                    "执行脚本"
                ],
                "parameters": [
                    {
                        "description": "ScriptService",
                        "name": "Script",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.ScriptService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/services.ScriptService"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "创建任务",
                "tags": [
                    "执行脚本"
                ],
                "parameters": [
                    {
                        "description": "ScriptService",
                        "name": "Script",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.ScriptService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/services.ScriptService"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/script/{script}": {
            "delete": {
                "description": "删除执行脚本",
                "tags": [
                    "执行脚本"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "script",
                        "name": "script",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "version",
                        "name": "version",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "返回码",
                    "type": "integer"
                },
                "data": {
                    "description": "返回数据"
                },
                "error": {
                    "description": "错误信息",
                    "type": "string"
                },
                "message": {
                    "description": "返回信息",
                    "type": "string"
                }
            }
        },
        "services.ScriptService": {
            "type": "object",
            "properties": {
                "kind": {
                    "$ref": "#/definitions/types.KindType"
                },
                "metadata": {
                    "$ref": "#/definitions/types.ScriptMetadata"
                },
                "source": {
                    "type": "string"
                }
            }
        },
...
        "types.KindType": {
            "type": "string",
            "enum": [
                "Script",
                "Task",
                "TaskTemplate"
            ],
            "x-enum-varnames": [
                "ScriptKind",
                "TaskKind",
                "TaskTemplateKind"
            ]
        },
        "types.ScriptMetadata": {
            "type": "object",
            "properties": {
                "createAt": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/types.ScriptType"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "types.ScriptType": {
            "type": "string",
            "enum": [
                "shell",
                "python"
            ],
            "x-enum-varnames": [
                "ScriptTypeShell",
                "ScriptTypePython"
            ]
        },
...
    }
}